Configure Iguana, write script to cat queries.
This PR introduces the configuration for the Iguana benchmarking tool.
It also includes code which is used to collect and concatenate all the
queries written and collected so far into a single file, one line per
query. This file is then fed into the Iguana benchmark tool as input.
Finally, the PR also covers the `npm run` commands used to launch
the benchmarking tool.

closes #24