Clarify benchmark issues, fix upload bug
This PR clarifies the need to call the benchmarking script from two
separate bash commands as opposed to using the postScriptHook in the
`config.yml` file in the benchmarking directory. It also adds comments
to the `config.mjs` file to remind future developers that changes to the
connection constants must also be updated in the `config.yml` file.

closes #38