Write script to automatically count triples (#36)
This PR introduces code that automates the counting of triples. It
counts the number of triples per Database per Graph. For example,
Virtuoso/MeSH is one instance of a count, whereas Neptune/Mesh is a
separate count. For the moment, part of the script is surrounded by
a try catch block to account for time outs from different endpoints.
This will likely need to be addressed if this script ends up being used
more frequently than at present.

closes #35