Run Virtuoso benchmarking
This PR documents the running of the Iguana benchmarking process solely
using our Virtuoso tech stack. It runs the benchmarks for 1 hour. The
results are then retrieved using a sparql query and parsed further
using python to produce a human readable table of results.

Certain parts of this process must be run manually. These are documented
in the `README.md` for at `src/benchmarking`.

closes 44